feat: 온보딩 화면 UI 수정
