+++ app/src/main/AndroidManifest.xml
@@ -28,6 +28,7 @@
 
         <activity android:name=".ui.main.MainActivity" />
         <activity android:name=".ui.login.LoginActivity" />
+        <activity android:name=".ui.login.onboarding.OnboardingActivity" />
         <activity
             android:name=".ui.join.BasicInformationActivity"
             android:windowSoftInputMode="adjustResize" />
