+++ app/src/main/res/layout/activity_onboarding.xml
@@ -4,32 +4,33 @@
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="match_parent"
         android:background="@color/gray_25"
-        android:paddingTop="32dp"
         android:paddingBottom="40dp">
+
+        <androidx.viewpager2.widget.ViewPager2
+            android:id="@+id/vp_onboarding"
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            app:layout_constraintLeft_toLeftOf="parent"
+            app:layout_constraintRight_toRightOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintBottom_toTopOf="@id/tv_next_button"/>
 
         <com.google.android.material.tabs.TabLayout
             android:id="@+id/tl_indicator"
             android:layout_width="wrap_content"
             android:layout_height="8dp"
             android:background="@color/white"
+            android:layout_marginBottom="152dp"
+            app:layout_constraintBottom_toTopOf="@id/tv_next_button"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
             app:tabBackground="@drawable/selector_dot_8"
             app:tabGravity="center"
             app:tabIndicatorHeight="0dp"
             app:tabMaxWidth="14dp"
             app:tabPaddingStart="6dp" />
-
-        <androidx.viewpager2.widget.ViewPager2
-            android:id="@+id/vp_onboarding"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tl_indicator" />
 
         <TextView
             android:id="@+id/tv_next_button"
@@ -46,6 +47,6 @@
             android:textColor="@color/white"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/vp_onboarding" />
+            app:layout_constraintBottom_toBottomOf="parent" />
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
